Se soluciono el error de que en descuento directo de ventas que si tenias mas de 2 productos no te pusiera el precio de solo 1

+++ PuntoDeVentaV2/AgregarDescuentoProducto.cs
@@ -2060,7 +2060,7 @@ namespace PuntoDeVentaV2
             {
                 if (!string.IsNullOrWhiteSpace(tb.Text))
                 {
-                    if (Convert.ToInt32(tb.Text)>=100)
+                    if (Convert.ToDecimal(tb.Text)>=100)
                     {
                         tb.Text = "99";
                     }
